Move find_peaks method
