+++ gammapy/estimators/utils.py
